+++ victreebot/extentions/interactions/locations/locations.py
@@ -241,7 +241,6 @@ async def command_locations(
         log_location_add,
         log_location_delete,
         log_location_edit,
-        log_location_info,
         log_location_list,
         *none,
     ) = await db.get_guild_log_settings(
@@ -251,7 +250,6 @@ async def command_locations(
             "log_location_add",
             "log_location_delete",
             "log_location_edit",
-            "log_location_info",
             "log_location_list",
         ],
     )
@@ -282,9 +280,6 @@ async def command_locations(
         .add_button(hikari.ButtonStyle.PRIMARY, "list_gyms")
         .set_label(SUPPORTED_LANGUAGES.get(language).location_action_row_list_gym)
         .add_to_container()
-        .add_button(hikari.ButtonStyle.PRIMARY, "info_gym")
-        .set_label(SUPPORTED_LANGUAGES.get(language).location_action_row_info_gym)
-        .add_to_container()
     )
     action_row_2 = (
         ctx.rest.build_action_row()
@@ -299,9 +294,6 @@ async def command_locations(
         .add_to_container()
         .add_button(hikari.ButtonStyle.PRIMARY, "list_pokestops")
         .set_label(SUPPORTED_LANGUAGES.get(language).location_action_row_list_pokestop)
-        .add_to_container()
-        .add_button(hikari.ButtonStyle.PRIMARY, "info_pokestop")
-        .set_label(SUPPORTED_LANGUAGES.get(language).location_action_row_info_pokestop)
         .add_to_container()
     )
 
@@ -475,59 +467,6 @@ async def command_locations(
                         await event.interaction.create_initial_response(
                             hikari.interactions.base_interactions.ResponseType.DEFERRED_MESSAGE_UPDATE, values[index]
                         )
-
-        elif event.interaction.custom_id == "info_gym" or event.interaction.custom_id == "info_pokestop":
-            location_type = "gym" if event.interaction.custom_id == "info_gym" else "pokéstop"
-            location_type = f"'{location_type}'"
-
-            location_name = await get_location_name(ctx, event, location_type, db, bot, bot_aware)
-            location_name = location_name.replace("'", "''")
-            location_name = f"'{location_name}'"
-            results = await db.get_location_info(ctx.get_guild(), location_type, location_name)
-            if results == []:
-                response = SUPPORTED_LANGUAGES.get(language).response_location_info_no_results.format(
-                    location_type=location_type.strip("'")
-                )
-                await ctx.create_followup(response, delete_after=auto_delete, embed=None, components=None)
-                if log_errors:
-                    log_response = SUPPORTED_LANGUAGES.get(language).log_response_location_info_no_results.format(
-                        datetime=await bot.get_timestamp_aware(gmt),
-                        member=ctx.member,
-                        location_type=location_type.strip("'"),
-                    )
-                    await bot.log_from_ctx(ctx, db, log_response)
-                return
-
-            known_coordinates = "[{google_maps_translation}](https://www.google.com/maps/place/{latitude},{longitude})"
-            unknown_coordinates = f"`{SUPPORTED_LANGUAGES.get(language).location_no_coordinates_set}`"
-
-            auto_delete = 20 if auto_delete < 20 else auto_delete
-            embed = (
-                hikari.Embed(
-                    title=SUPPORTED_LANGUAGES.get(language).location_info_embed_title,
-                    description=f"""`{location_type.capitalize().strip("'")}:` {location_name.capitalize().strip("'")}\n `Description:` {results[0].get('description') if results[0].get('description') else SUPPORTED_LANGUAGES.get(language).location_no_description_set}""",  # noqa E501
-                    colour=hikari.Colour(0x8BC683),
-                )
-                .add_field(
-                    name=SUPPORTED_LANGUAGES.get(language).location_info_embed_field_coordinates,
-                    value=f"Latitude: {'`'+str(results[0].get('latitude'))+'`' if results[0].get('latitude') else unknown_coordinates}\n Longitude: {'`'+str(results[0].get('longitude'))+'`' if results[0].get('longitude') else unknown_coordinates}",  # noqa E501
-                    inline=False,
-                )
-                .add_field(
-                    name=SUPPORTED_LANGUAGES.get(language).location_info_embed_field_google_maps,
-                    value=f"{known_coordinates.format(google_maps_translation=SUPPORTED_LANGUAGES.get(language).location_google_maps, latitude=str(results[0].get('latitude')), longitude=str(results[0].get('longitude'))) if results[0].get('latitude') is not None and results[0].get('longitude') is not None else unknown_coordinates}",  # noqa E501
-                    inline=False,
-                )
-            )
-
-            await ctx.create_followup(embed=embed, delete_after=auto_delete, components=None)
-            if log_location_info:
-                log_response = SUPPORTED_LANGUAGES.get(language).log_response_location_info_requested.format(
-                    datetime=await bot.get_timestamp_aware(gmt),
-                    member=ctx.member,
-                    location_type=location_type.strip("'"),
-                )
-                await bot.log_from_ctx(ctx, db, log_response)
 
         elif event.interaction.custom_id == "add_gym" or event.interaction.custom_id == "add_pokestop":
             location_type = "gym" if event.interaction.custom_id == "add_gym" else "pokéstop"